+++ smos-docs-site/content/pages/changelog.markdown
@@ -18,6 +18,11 @@ description: The changelog for all of the Smos tools and libraries
   You can now exit any report using `Esc`, even when no file had been selected beforehand.
   Thank you [@ketzacoatl](https://github.com/ketzacoatl)!
 
+## Changed
+
+- `smos-scheduler`: Explicitly setting a header's `state` to `null` now leaves the generated state as null instead of setting it to `TODO`.
+
+
 # <a name="2021-02-01">[2021-02-01](#2021-02-01)
 - <a name="smos-0.1.3">[smos 0.1.3](#smos-0.1.3)
 - <a name="smos-query-0.1.2">[smos-query 0.1.2](#smos-query-0.1.2)